pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "questmaster"
version = "1.1.0"
description = "Club JDR booking app"
requires-python = ">=3.13"
license = "MIT"
dependencies = [
    "Flask==2.3.3",
    "Flask-Discord==0.1.69",
    "Flask-Migrate==4.1.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-WTF==1.2.2",
    "Flask-Caching==2.3.1",
    "Flask-Admin==1.6.1",
    "redis==5.0.8",
    "gunicorn==23.0.0",
    "psutil==5.9.8",
    "psycopg2-binary==2.9.11",
    "Werkzeug==3.1.5",
    "pyyaml==6.0.3",
    "schema==0.7.8",
    "tenacity==8.2.3",
    "wtforms==3.1.2",
    "python-dateutil==2.9.0.post0",
    "unidecode==1.4.0",
    "python-slugify==8.0.4",
    "apscheduler==3.11.2",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.2",
    "pytest-cov==5.0.0",
    "python-dotenv==1.2.1",
    "beautifulsoup4==4.14.3",
    "blinker==1.9.0",
]
lint = [
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests requiring live external services (Discord API)",
]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99
known_first_party = ["website", "config", "tests"]

[tool.setuptools.packages.find]
include = ["website*"]
